fix(via): change default max connections to 1000
This gives the runtime a bit of breathing room. It's probably unnecessary on Linux but definitely required on a Mac. If this is truly specific to macOS, we can make a cfg for macOS. That can happen in a follow up when I have time to setup and test on a VM.