feat(via): accept loop improvements (#219)
* feat(via): accept loop improvements

Solidify a garbage collection and connection management strategy.

The idea is try to make progress on an inflight connection while we wait for a new connection every other turn of the accept loop.

If we end up accepting a new connection, clean up any connection in the `JoinSet` that we know are finished.

* fix(via): Add JoinError variant to ServerError enum

* fix(via): change default max connections to 1000

This gives the runtime a bit of breathing room. It's probably unnecessary on Linux but definitely required on a Mac. If this is truly specific to macOS, we can make a cfg for macOS. That can happen in a follow up when I have time to setup and test on a VM.

* fix(via): join max of 2 connections